% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append-colors.R
\name{appendColors}
\alias{appendColors}
\title{Add colors to the data}
\usage{
appendColors(.data, dimension_colors)
}
\arguments{
\item{.data}{}

\item{dimension_colors}{color table}
}
\description{
Add colors to the data
}
