% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-data.R
\name{compare_old_and_new_data}
\alias{compare_old_and_new_data}
\title{Title}
\usage{
compare_old_and_new_data(
  old_dat_path,
  new_dat_path,
  compare_var,
  raw_or_inequal = "raw",
  cutoff_for_odd = 1,
  doplot = TRUE,
  alphaval = 0.2,
  show_example = TRUE
)
}
\arguments{
\item{show_example}{}
}
\description{
Title
}
