% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-non-spatial-data.R
\name{create_non_spatial_data}
\alias{create_non_spatial_data}
\title{Title}
\usage{
create_non_spatial_data(
  data_heat_raw,
  data_countries,
  dimension_colors,
  dataset_name,
  development = FALSE
)
}
\arguments{
\item{data_heat_raw}{}

\item{data_countries}{}

\item{dimension_colors}{}

\item{development}{}
}
\description{
Title
}
