% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{duplicate_survey_geo}
\alias{duplicate_survey_geo}
\title{Pre-built country/years where the geography for a survey should get duplicated
so that it can be used for more than one year}
\usage{
duplicate_survey_geo(.data)
}
\arguments{
\item{.data}{}
}
\value{

}
\description{
Pre-built country/years where the geography for a survey should get duplicated
so that it can be used for more than one year
}
