% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-tables.R
\name{HEAT_create_random_sample}
\alias{HEAT_create_random_sample}
\title{Title}
\usage{
HEAT_create_random_sample(
  .data = heatdata::data_heat_raw,
  use_heatdata_strata = TRUE,
  setting_choice = NULL,
  dimension_choice = NULL,
  indicator_choice = NULL,
  n_per_dimension = NULL,
  intentional_errors_to_add = NULL
)
}
\arguments{
\item{n_per_dimension}{}
}
\description{
Title
}
