% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-tables.R
\name{HEAT_create_setyrsrc_table}
\alias{HEAT_create_setyrsrc_table}
\title{Create a table with the unique set of setting, year and source}
\usage{
HEAT_create_setyrsrc_table(.data)
}
\arguments{
\item{.data}{}
}
\description{
the actual names of the setting, year and source variables (which
might be setting, year and source) are in the heatmeasures::setting_year_source_variables
vector
}
