% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-tables.R
\name{HEAT_create_strata_table}
\alias{HEAT_create_strata_table}
\title{Create the strata table (unique set of strata)}
\usage{
HEAT_create_strata_table(.data)
}
\arguments{
\item{.data}{}
}
\value{

}
\description{
this function makes use of the heatmeasures::strata_variables for the
definition of what is a strata
}
