% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-tables.R
\name{HEAT_create_subregionminmax_table}
\alias{HEAT_create_subregionminmax_table}
\title{Compute the table of subregion min and max values for the mapping (only required for HEAT)}
\usage{
HEAT_create_subregionminmax_table(.data)
}
\arguments{
\item{.data}{}
}
\description{
Compute the table of subregion min and max values for the mapping (only required for HEAT)
}
