% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-add-variables.R
\name{HEAT_data_add_variables}
\alias{HEAT_data_add_variables}
\title{Add required new variables to the dataset}
\usage{
HEAT_data_add_variables(.data, table_type = "all")
}
\arguments{
\item{.data}{input data}

\item{table_type}{either "who_data" or "country_data"}
}
\description{
Add required new variables to the dataset
}
