% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-fixes.R
\name{HEAT_data_fixes}
\alias{HEAT_data_fixes}
\title{Apply specific data fixes to raw data}
\usage{
HEAT_data_fixes(.data, table_type = "all")
}
\arguments{
\item{table_type}{"country_data", "heat_data" (internal database)}
}
\value{

}
\description{
These might include fixes that need to be applied to all data (remove rows
with all NA values) or fixes that should
}
