% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-fixes.R
\name{HEAT_drop_defective_strata}
\alias{HEAT_drop_defective_strata}
\title{Drop strata with all missing estimates}
\usage{
HEAT_drop_defective_strata(.data)
}
\arguments{
\item{.data}{}
}
\value{

}
\description{
Drop strata with all missing estimates
}
