% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-fixes.R
\name{HEAT_force_variable_types}
\alias{HEAT_force_variable_types}
\title{Force variables to be the types they're supposed to be}
\usage{
HEAT_force_variable_types(.data, table_type = "all")
}
\arguments{
\item{.data}{}
}
\description{
Force variables to be the types they're supposed to be
}
