% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-fixes.R
\name{HEAT_rename_variables}
\alias{HEAT_rename_variables}
\title{Rename a table's variables}
\usage{
HEAT_rename_variables(.data, table_type = "all")
}
\arguments{
\item{table_type}{}
}
\description{
Occasionally the database we've received has a slightly different name than the app expects
and we need to alter it to fit.
}
