% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{HEAT_table_validation_tests}
\alias{HEAT_table_validation_tests}
\title{Apply validation tests to the table}
\usage{
HEAT_table_validation_tests(.data, tests_to_run)
}
\arguments{
\item{tests_to_run}{}
}
\description{
Apply validation tests to the table
}
\examples{
var_names <- names(heatdata::HEAT_variable_descriptions)
var_tests <- stringr::str_extract(var_names, "^test_.*")
var_results <- na.exclude(var_tests)

# test_results <- HEAT_table_validation_tests(data_heat_raw, tests)
}
