% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatdata.R
\docType{package}
\name{heatdata-package}
\alias{heatdata}
\alias{heatdata-package}
\title{HEAT Data application}
\description{
An R package to generate data for the Health Equity Assessment Toolkit.
}
\author{
\strong{Maintainer}:  ZevRoss Spatial Anaylsis \email{info@zevross.com}

Other contributors:
\itemize{
  \item Zev Ross [contributor]
}

}
