% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{multi_use_survey_geo}
\alias{multi_use_survey_geo}
\title{Add new record in DHS survey geo dataset that is a duplicate of an existing year except includes a new year.}
\usage{
multi_use_survey_geo(
  .data,
  setting_iso3,
  existing_year,
  new_years,
  keep_existing = TRUE
)
}
\arguments{
\item{setting_iso3}{ISO3 of focus country}

\item{existing_year}{the year with real data from the DHS survey}

\item{new_years}{the years you want to get the same data as the existing year}

\item{keep_existing}{should the record for the existing year stay in the dataset}
}
\value{

}
\description{
Add new record in DHS survey geo dataset that is a duplicate of an existing year except includes a new year.
}
\details{
as an example, the DHS survey might include Peru 2007 but the WHO HEAT data includes Peru 2008 and 2008
is not available in the DHS data. Use this function to add a new recode with year 2008 that is identical
to 2007 except the year is changed to 2008. The default is to keep the existing record so in this case you
end up with two records with the same geography but one has year = 2007 and one is year = 2008.
}
