% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{spatial_fix_survey_years}
\alias{spatial_fix_survey_years}
\title{Change the DHS shapefile year so that the year matches the
WHO HEAT year -- for DHS surveys that spanned two years, DHS
may have assigned one year while WHO HEAT assigned the other year}
\usage{
spatial_fix_survey_years(.data)
}
\arguments{
\item{.data}{}
}
\description{
Change the DHS shapefile year so that the year matches the
WHO HEAT year -- for DHS surveys that spanned two years, DHS
may have assigned one year while WHO HEAT assigned the other year
}
