% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{survey_boundary_download}
\alias{survey_boundary_download}
\title{survey_boundary_download}
\usage{
survey_boundary_download(sel, link)
}
\arguments{
\item{link}{}
}
\value{

}
\description{
survey_boundary_download
}
