% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{survey_boundary_move_unzip}
\alias{survey_boundary_move_unzip}
\title{Title}
\usage{
survey_boundary_move_unzip(
  newpath,
  survey,
  download_folder = "~/Downloads",
  search_pattern = "**/sdr_*.zip"
)
}
\arguments{
\item{search_pattern}{}
}
\description{
Title
}
