% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{survey_boundary_prep}
\alias{survey_boundary_prep}
\title{Title}
\usage{
survey_boundary_prep(
  sel,
  country_id,
  spatialdir,
  download_folder = "~/Downloads"
)
}
\arguments{
\item{sel}{a selenium connection sel <- rsDriver(port = 3000L, browser = "chrome")}

\item{spatialdir}{}
}
\value{

}
\description{
Title
}
