% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-tests.R
\name{test_has_required_variables}
\alias{test_has_required_variables}
\title{Test to confirm that all required variables are present in the dataset}
\usage{
test_has_required_variables(.data)
}
\arguments{
\item{.data}{}
}
\description{
required variables are listed originally in an Excel file in data-raw
but were processed into a variable exposed by this package (\code{HEAT_variable_descriptions})
}
