% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{union_subnational_boundaries}
\alias{union_subnational_boundaries}
\title{Title}
\usage{
union_subnational_boundaries(.data, just_testing = FALSE)
}
\arguments{
\item{just_testing}{}
}
\description{
Title
}
